Move `DisplayFolderRepository` to `:legacy:ui:folder` module
