+++ legacy/ui/folder/src/main/java/app/k9mail/legacy/ui/folder/KoinModule.kt
@@ -1,0 +1,13 @@
+package app.k9mail.legacy.ui.folder
+
+import org.koin.dsl.module
+
+val uiFolderModule = module {
+    single {
+        DisplayFolderRepository(
+            accountManager = get(),
+            messagingController = get(),
+            messageStoreManager = get(),
+        )
+    }
+}
