+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/managefolders/ManageFoldersViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.folder.DisplayFolder
-import app.k9mail.legacy.mailstore.DisplayFolderRepository
+import app.k9mail.legacy.ui.folder.DisplayFolderRepository
 
 class ManageFoldersViewModel(
     private val folderRepository: DisplayFolderRepository,
